Make use of the flexibility of online meetings and schedule some important meetings in the week before?
This is a PR to discuss opportunities for more flexible scheduling in order to reduce conflicts during the plenary meeting week as this topic came up during the IETF-111 discussion. This PR does not propose to simply use more meeting days (with less parallel sessions) but maybe schedule different kind of sessions differently. This is just some input for consideration for future online meetings and not necessarily a fully fleshed out proposal, however, so is also the rest of the doc as these are more guidelines and experiences rather than fixed rules.